split propagation control handlers
